// src/api/models/BankAccountModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type BankAccountModel = {
  businessId?: string;
  accountName?: string | null;
  accountNumber?: string | null;
  bankCode?: string | null;
  bankName?: string | null;
};