// src/api/models/BusinessKYCModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type BusinessKYCModel = {
  businessId?: string;
  businessCapacity?: string | null;
  officeAddress?: string | null;
  cacDocumentReference?: string | null;
  meansOfIdentification?: string | null;
};