// src/api/models/SavedBusinessViewStandardResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { HttpStatusCode } from './HttpStatusCode';
import type { SavedBusinessView } from './SavedBusinessView';
export type SavedBusinessViewStandardResponse = {
  href?: string | null;
  relations?: Array<string> | null;
  method?: string | null;
  routeName?: string | null;
  routeValues?: any;
  status?: boolean;
  message?: string | null;
  data?: SavedBusinessView;
  statusCode?: HttpStatusCode;
  errors?: any;
};