// src/api/models/SavedServiceViewPagedCollectionStandardResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { HttpStatusCode } from './HttpStatusCode';
import type { SavedServiceViewPagedCollection } from './SavedServiceViewPagedCollection';
export type SavedServiceViewPagedCollectionStandardResponse = {
  href?: string | null;
  relations?: Array<string> | null;
  method?: string | null;
  routeName?: string | null;
  routeValues?: any;
  status?: boolean;
  message?: string | null;
  data?: SavedServiceViewPagedCollection;
  statusCode?: HttpStatusCode;
  errors?: any;
};